requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
Flask-Mail
face_recognition
opencv-python
numpy
pillow
